🚧 9RCWZQ task: apply external agent result
