+++ .agentplane/tasks/202608252330-9RCWZQ/pr/diffstat.txt
@@ -1,7 +1,8 @@
- .../cli/run-cli.core.pr-flow.pr-open.git.test.ts   | 229 +++++++++++++++++++--
- .../run-cli.core.pr-flow.pr-open.network.test.ts   |  79 ++++++-
+ .../cli/run-cli.core.pr-flow.pr-open.git.test.ts   |  35 ++--
+ .../run-cli.core.pr-flow.pr-open.network.test.ts   |  79 +++++++-
  .../src/commands/pr/internal/git-host-identity.ts  |   2 +-
- .../src/commands/pr/internal/provider-base.ts      | 111 ++++++++++
- .../src/commands/pr/internal/sync-github.test.ts   |  44 ++++
+ .../src/commands/pr/internal/provider-base.test.ts | 202 +++++++++++++++++++++
+ .../src/commands/pr/internal/provider-base.ts      | 111 +++++++++++
+ .../src/commands/pr/internal/sync-github.test.ts   |  44 +++++
  .../agentplane/src/commands/pr/internal/sync.ts    |  19 +-
- 6 files changed, 457 insertions(+), 27 deletions(-)
+ 7 files changed, 465 insertions(+), 27 deletions(-)
