+++ .agentplane/tasks/202608252330-9RCWZQ/pr/meta.json
@@ -2,18 +2,11 @@
   "base": "main",
   "branch": "task/202608252330-9RCWZQ/allow-exact-sha-release-tasks-to-open-hosted-prs",
   "created_at": "2026-08-26T02:00:01.912Z",
-  "diffstat_sha256": "sha256:0c037f9f15dd76669f4c85847cb45a2aa3e90c04c98814a16095bc0ca10ed12f",
-  "last_verified_at": "2026-08-26T08:49:52.154Z",
-  "last_verified_diffstat_sha256": "sha256:3f53ba00150f7747b5d09ff30f347357d9f666a06f399202b8bacc626b423d65",
+  "diffstat_sha256": "sha256:c2073ba7aa1ba0eee2031724911898532c0bc407580c6b2dabb1ca77196b70ec",
+  "last_verified_at": "2026-08-26T09:25:06.899Z",
+  "last_verified_diffstat_sha256": "sha256:0c037f9f15dd76669f4c85847cb45a2aa3e90c04c98814a16095bc0ca10ed12f",
   "pr_number": 4889,
   "pr_url": "https://github.com/basilisk-labs/agentplane/pull/4889",
-  "pre_merge_closure": {
-    "basis_commit": "c2547c753008afa4a35a924b0c1b0e6128781bd1",
-    "branch": "task/202608252330-9RCWZQ/allow-exact-sha-release-tasks-to-open-hosted-prs",
-    "pr_number": 4889,
-    "recorded_at": "2026-08-26T08:52:04.319Z",
-    "state": "closed_before_merge"
-  },
   "provider": {
     "hostname": "github.com",
     "kind": "github",
@@ -27,6 +20,6 @@
   "task_id": "202608252330-9RCWZQ",
   "updated_at": "2026-08-26T02:46:09.863Z",
   "verify": {
-    "status": "pass"
+    "status": "fail"
   }
 }
