+++ .agentplane/tasks/202608252330-9RCWZQ/pr/review.md
@@ -6,14 +6,14 @@ Created: 2026-08-26T02:00:01.912Z
 
 - Task: `202608252330-9RCWZQ`
 - Title: Allow exact-SHA release tasks to open hosted PRs against the matching protected base branch
-- Status: DONE
+- Status: DOING
 - Branch: `task/202608252330-9RCWZQ/allow-exact-sha-release-tasks-to-open-hosted-prs`
 - Canonical task record: `.agentplane/tasks/202608252330-9RCWZQ/README.md`
 
 ## Verification
 
-- State: ok
-- Note: Verified after recovery: exact-SHA provider-base implementation is unchanged, exact-head provider payload tests pass 19/19, focused PR-open tests and typecheck pass, and bounded full regression evidence is passing.
+- State: needs_rework
+- Note: Rework: Declared check failed: bun run ci:local:full
 - Canonical workflow state lives in the task README.
 
 ## Handoff Notes
@@ -29,13 +29,14 @@ Created: 2026-08-26T02:00:01.912Z
 - Head: computed live by `agentplane pr check` / `agentplane integrate`
 
 ```text
- .../cli/run-cli.core.pr-flow.pr-open.git.test.ts   | 229 +++++++++++++++++++--
- .../run-cli.core.pr-flow.pr-open.network.test.ts   |  79 ++++++-
+ .../cli/run-cli.core.pr-flow.pr-open.git.test.ts   |  35 ++--
+ .../run-cli.core.pr-flow.pr-open.network.test.ts   |  79 +++++++-
  .../src/commands/pr/internal/git-host-identity.ts  |   2 +-
- .../src/commands/pr/internal/provider-base.ts      | 111 ++++++++++
- .../src/commands/pr/internal/sync-github.test.ts   |  44 ++++
+ .../src/commands/pr/internal/provider-base.test.ts | 202 +++++++++++++++++++++
+ .../src/commands/pr/internal/provider-base.ts      | 111 +++++++++++
+ .../src/commands/pr/internal/sync-github.test.ts   |  44 +++++
  .../agentplane/src/commands/pr/internal/sync.ts    |  19 +-
- 6 files changed, 457 insertions(+), 27 deletions(-)
+ 7 files changed, 465 insertions(+), 27 deletions(-)
 ```
 
 </details>
